+++ R/diffQced.R
@@ -35,11 +35,23 @@ diffQced <- function(.file, .side_by_side = TRUE, .ignore_white_space = FALSE){
     cli::cli_abort(glue::glue("'{.file}' has no record in the QC log"))
   }
   
-  file_info <- svnLog(.file) %>% dplyr::slice(1)
+  file_log <- svnLog(.file)
+  file_info <- file_log %>% dplyr::slice(1)
+  
+  authors_last_qc <-
+    file_log %>% 
+    dplyr::filter(rev > qced_revision) %>% 
+    dplyr::pull(author)
   
   cli::cli_h2(glue::glue("QC diff for: ", .file))
   cli::cli_inform(glue::glue("Last QCed Revision: ", qced_revision))
   cli::cli_inform(glue::glue("Last Author: ", file_info$author))
+  
+  this_user <- Sys.info()[["user"]]
+  
+  if (this_user %in% authors_last_qc) {
+    cli::cli_alert_warning(glue::glue("'{this_user}' has modified '{basename(.file)}' since last QC"))
+  }
   
   diffPreviousRevisions(
     .file = .file,