+++ NEWS.md
@@ -1,3 +1,9 @@
+# review development
+
+## New features and changes
+
+- Added warning message to `diffQCed` when user has modified file since last QC. (#61)
+
 # review 3.7.0
 
 ## New features and changes